Finished main game loop
Completed full game loop; next iteration of work done to "complete" this game would be to add audio and a settings menu for the volume of the music/sfx.

I won't do graphics updates on this after I do audio, simply because custom visuals in CustomTkinter is not as efficient as Android Studio or Godot, for example.

However, I would say that with this final update, complete with a default 10 question quiz on 4 elementary operations for 1, 2, and 3 digits, randomized, the game is "playable".